Bugfix: dont order deletion for still-not-started sessions
git-svn-id: file:///home/code/svn/ovd/branches/2.5@3015 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/check.php
@@ -54,7 +54,7 @@ if ($sessions) {
 	foreach ($sessions as $session) {
 		$buf = $session->getStatus();
 
-		if (! $buf || (int)$buf == 4) {
+		if ((! $buf || (int)$buf == 4) && $session->getAttribute('start_time') < (time()-10)) {
 			Logger::warning('main', '(check) Session \''.$session->id.'\' is no longer existing, deleting');
 
 			Abstract_Session::delete($session->id);
